fix running tester form process

+++ LogExplorer/Services/Helpers/FileHelper.cs
@@ -270,16 +270,9 @@ namespace LogExplorer.Services.Helpers
             {
                 WindowStyle = hidden ? ProcessWindowStyle.Hidden : ProcessWindowStyle.Normal,
                 WorkingDirectory = Path.GetDirectoryName(filePath),
-                FileName = "cmd.exe",
-                Arguments = $@"/C {GetFileName(filePath, true)} {arguments}"
+                FileName = filePath,
+                Arguments = arguments
             };
-            //todo fix this way
-            //var startInfo = new ProcessStartInfo
-            //{
-            //    WindowStyle = hidden ? ProcessWindowStyle.Hidden : ProcessWindowStyle.Normal,
-            //    FileName = filePath,
-            //    Arguments = arguments
-            //       };
             process.StartInfo = startInfo;
 			var processStr = $"'{GetFileName(filePath, true)} {arguments}'";
 			try
@@ -294,15 +287,10 @@ namespace LogExplorer.Services.Helpers
 			}
 			catch (OperationCanceledException)
 			{
-#region todo delete when fixed 
-                var processlist = Process.GetProcesses().Where(p=>p.ProcessName.Contains(GetFileName(filePath)));
-
-                foreach (var coProcess in processlist)
-			    {
-			        coProcess.Kill();
-			    }
-#endregion
-
+                if (process.HasExited)
+                {
+                    throw;
+                }
                 process.Kill();
 				Logger.AddMessage(Messages.KillingProccess);
 				throw;
